Pre-Public Push
Before making the repository public, needed to re-think my original plans. If the code was only to be between myself and the individual(s) reviewing the code, I could stick with the original plan. However on setting to public, need to resort to consider security.

Minor changes include improving error message responses.